# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Game API",
        "version": "1.1.0",
        "description": "Simple game api that store user account, biodata, and game histories"
    },
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer [JWT TOKEN]"
        }
    },
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "basePath": "/api/v1",
...
    "paths": {
...
        "/user_games/histories": {
            "get": {
                "summary": "Get all user game histories",
                "operationId": "GetUserGameHistories",
                "responses": {
                    "200": {
                        "description": "Returns a list of user game histories",
                        "schema": {
                            "$ref": "#/responses/getUserGameHistoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game History"
                ]
            },
            "post": {
                "summary": "Create a new user game history",
                "operationId": "CreateUserGameHistories",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns user game history that has been created",
                        "schema": {
                            "$ref": "#/responses/postUserGameHistoriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game History"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "publisher",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "cover",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "score",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1,
                        "maximum": 100
                    },
                    {
                        "name": "userGameId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                ]
            }
        },
        "/user_game/history/{id}": {
            "get": {
                "summary": "Get user game history by id",
                "operationId": "GetUserGameHistory",
                "responses": {
                    "200": {
                        "description": "Returns a user game history",
                        "schema": {
                            "$ref": "#/responses/getUserGameHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "404": {
                        "description": "User game history not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game history id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "User Game History"
                ]
            },
            "patch": {
                "summary": "Update user game history by id",
                "operationId": "UpdateUserGameHistory",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns before after user game history field that has been updated",
                        "schema": {
                            "$ref": "#/responses/patchUserGameHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game history not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game History"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game history id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "publisher",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "cover",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "score",
                        "in": "formData",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1,
                        "maximum": 100
                    },
                    {
                        "name": "userGameId",
                        "in": "formData",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                ]
            },
            "delete": {
                "summary": "Delete user game history by id",
                "operationId": "DeleteUserGameHistory",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user game history that has been deleted",
                        "schema": {
                            "$ref": "#/responses/deleteUserGameHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game history not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game History"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game history id",
                        "required": true,
                        "type": "integer"
                    }
                ]
            }
        },
...
    },
...
    "responses": {
...
        "getUserGameHistoriesResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "count": {
                    "type": "integer",
                    "format": "int32"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameHistoryInclude"
                    }
                }
            }
        },
        "postUserGameHistoriesResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 201
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameHistory created successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGameHistory"
                }
            }
        },
        "getUserGameHistoryResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "data": {
                    "$ref": "#/definitions/userGameHistoryInclude"
                }
            }
        },
        "patchUserGameHistoryResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameHistory updated successfully"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "before": {
                            "$ref": "#/requestBodies/patchUserGameHistoryRequest"
                        },
                        "after": {
                            "$ref": "#/requestBodies/patchUserGameHistoryRequest"
                        }
                    }
                }
            }
        },
        "deleteUserGameHistoryResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGameHistory deleted successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGameHistory"
                }
            }
        },
...
        "generalErrorResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "badRequestResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 400
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "value": {
                                "type": "string",
                                "format": "string"
                            },
                            "msg": {
                                "type": "string",
                                "format": "string"
                            },
                            "param": {
                                "type": "string",
                                "format": "string"
                            },
                            "location": {
                                "type": "string",
                                "format": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "userGame": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string",
                    "format": "string"
                },
                "password": {
                    "type": "string",
                    "format": "string"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "userGameHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "type": "string",
                    "format": "string"
                },
                "publisher": {
                    "type": "string",
                    "format": "string"
                },
                "cover": {
                    "type": "string",
                    "format": "string"
                },
                "lastPlayed": {
                    "type": "string",
                    "format": "date-time"
                },
                "score": {
                    "type": "number",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "userGameHistoryInclude": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "type": "string",
                    "format": "string"
                },
                "publisher": {
                    "type": "string",
                    "format": "string"
                },
                "cover": {
                    "type": "string",
                    "format": "string"
                },
                "lastPlayed": {
                    "type": "string",
                    "format": "date-time"
                },
                "score": {
                    "type": "number",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                },
                "UserGame": {
                    "$ref": "#/definitions/userGame"
                }
            }
        }
    }
}